test: split v2 uat registry contract tests
